+++ loader/app/index.ts
@@ -25,6 +25,7 @@ export {};
 
 if(__build.target === "client") {
     /* do this so we don't get a react dev tools warning within the client */
-    (window as any).__REACT_DEVTOOLS_GLOBAL_HOOK__ = (window as any).__REACT_DEVTOOLS_GLOBAL_HOOK__ || {};
-    (window as any).__REACT_DEVTOOLS_GLOBAL_HOOK__.inject = function () {};
+    if(!('__REACT_DEVTOOLS_GLOBAL_HOOK__' in window))
+        window.__REACT_DEVTOOLS_GLOBAL_HOOK__ = {};
+    window.__REACT_DEVTOOLS_GLOBAL_HOOK__.inject = function () {};
 }