+++ shared/backend.d/dns.d.ts
@@ -3,3 +3,4 @@ import {ServerAddress} from "tc-shared/ui/server";
 
 export function supported();
 export function resolve_address(address: ServerAddress, options?: ResolveOptions) : Promise<AddressTarget>;
+export function resolve_address_ipv4(address: string) : Promise<string>;